Navbar and first Login page
So the navbar works essentially; the destinations are still sort of missing and for the login the logic is missing but the view and states are there.
Also have to figure out whether to generate the token in the frontend or if we make a request only and catch it. Going to tackle that tomorrow